+++ lib/create_location/create_location_view.dart
@@ -71,7 +71,12 @@ class CreateLocationWidget extends StatelessWidget {
           bloc: BlocProvider.of<CreateLocationBloc>(context),
           builder: (context, CreateLocationState state) {
             if (state.loading) {
-              return Center(child: CircularProgressIndicator(strokeWidth: 4.0));
+              return Center(
+                  child: CircularProgressIndicator(
+                strokeWidth: 4.0,
+                valueColor: AlwaysStoppedAnimation<Color>(
+                    Theme.of(context).accentColor),
+              ));
             } else {
               return Center(
                 child: ListView(
@@ -200,6 +205,8 @@ class CreateLocationWidget extends StatelessWidget {
                     child: Center(
                         child: CircularProgressIndicator(
                       strokeWidth: 4.0,
+                      valueColor: AlwaysStoppedAnimation<Color>(
+                          Theme.of(context).accentColor),
                     ))),
                 errorWidget: (context, url, error) => Icon(Icons.error),
               )));