+++ lib/stats/charts/attempts_by_value/attempts_by_value.dart
@@ -44,7 +44,8 @@ class _AttemptsByValueChartState extends State<AttemptsByValueChart> {
   @override
   void initState() {
     filteredAttemptsStream = StreamController<List<Attempt>>.broadcast();
-    filteredAttemptsListener = filteredAttemptsStream.stream.listen((filteredAttempts) {
+    filteredAttemptsListener =
+        filteredAttemptsStream.stream.listen((filteredAttempts) {
       setState(() {
         chartSeries = _buildChartSeries(context, filteredAttempts);
       });
@@ -78,53 +79,81 @@ class _AttemptsByValueChartState extends State<AttemptsByValueChart> {
             padding: EdgeInsets.all(UIConstants.SMALLER_PADDING),
             decoration: BoxDecoration(
                 color: Theme.of(context).cardColor,
-                borderRadius: BorderRadius.all(Radius.circular(UIConstants.CARD_BORDER_RADIUS))),
+                borderRadius: BorderRadius.all(
+                    Radius.circular(UIConstants.CARD_BORDER_RADIUS))),
             child: (chartSeries != null)
                 ? charts.BarChart(
                     chartSeries,
                     primaryMeasureAxis: charts.NumericAxisSpec(
                         renderSpec: charts.GridlineRendererSpec(
                       labelStyle: charts.TextStyleSpec(
-                        fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
-                        fontWeight: Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
-                        color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
+                        fontSize: Theme.of(context)
+                            .accentTextTheme
+                            .caption
+                            .fontSize
+                            .toInt(),
+                        fontWeight: Theme.of(context)
+                            .accentTextTheme
+                            .caption
+                            .fontWeight
+                            .toString(),
+                        color: charts.ColorUtil.fromDartColor(
+                            Theme.of(context).accentColor),
                       ),
                       lineStyle: charts.LineStyleSpec(
-                        color:
-                            charts.ColorUtil.fromDartColor(Theme.of(context).dialogBackgroundColor),
+                        color: charts.ColorUtil.fromDartColor(
+                            Theme.of(context).dialogBackgroundColor),
                       ),
                     )),
                     domainAxis: charts.OrdinalAxisSpec(
                       renderSpec: charts.SmallTickRendererSpec(
                         lineStyle: charts.LineStyleSpec(
-                          color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
+                          color: charts.ColorUtil.fromDartColor(
+                              Theme.of(context).accentColor),
                         ),
                         labelStyle: charts.TextStyleSpec(
-                          fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
-                          fontWeight:
-                              Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
-                          color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
+                          fontSize: Theme.of(context)
+                              .accentTextTheme
+                              .caption
+                              .fontSize
+                              .toInt(),
+                          fontWeight: Theme.of(context)
+                              .accentTextTheme
+                              .caption
+                              .fontWeight
+                              .toString(),
+                          color: charts.ColorUtil.fromDartColor(
+                              Theme.of(context).accentColor),
                         ),
                         labelRotation: widget.rotateLabels ?? false ? 45 : 0,
                       ),
-                      tickProviderSpec: charts.StaticOrdinalTickProviderSpec(ticks),
+                      tickProviderSpec:
+                          charts.StaticOrdinalTickProviderSpec(ticks),
                     ),
                     barGroupingType: charts.BarGroupingType.stacked,
                     defaultInteractions: false,
                     behaviors: [
                       charts.SeriesLegend(
                         entryTextStyle: charts.TextStyleSpec(
-                          fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
-                          fontWeight:
-                              Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
-                          color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
+                          fontSize: Theme.of(context)
+                              .accentTextTheme
+                              .caption
+                              .fontSize
+                              .toInt(),
+                          fontWeight: Theme.of(context)
+                              .accentTextTheme
+                              .caption
+                              .fontWeight
+                              .toString(),
+                          color: charts.ColorUtil.fromDartColor(
+                              Theme.of(context).accentColor),
                         ),
                       ),
                     ],
                   )
                 : Center(
                     child: Text(
-                    "There are no existing attempts ${widget.attempts.isNotEmpty ? "matching these filters" : ""}. \nGo log some!",
+                    "There are no existing attempts${widget.attempts.isNotEmpty ? " matching these filters" : ""}. \nGo log some!",
                     style: Theme.of(context).accentTextTheme.subtitle2,
                     textAlign: TextAlign.center,
                   ))));
@@ -136,7 +165,8 @@ class _AttemptsByValueChartState extends State<AttemptsByValueChart> {
       return null;
     }
 
-    Map<String, Map<String, int>> valuesBySendType = <String, Map<String, int>>{};
+    Map<String, Map<String, int>> valuesBySendType =
+        <String, Map<String, int>>{};
     for (String sendType in SendTypes.SEND_TYPES) {
       valuesBySendType.putIfAbsent(sendType, widget.createEmptyMap);
     }
@@ -148,12 +178,14 @@ class _AttemptsByValueChartState extends State<AttemptsByValueChart> {
       }
     }
 
-    Map<String, List<AttemptsByValueSeries>> chartDataMap = <String, List<AttemptsByValueSeries>>{};
+    Map<String, List<AttemptsByValueSeries>> chartDataMap =
+        <String, List<AttemptsByValueSeries>>{};
 
     for (String sendType in SendTypes.SEND_TYPES) {
       chartDataMap.putIfAbsent(sendType, () => <AttemptsByValueSeries>[]);
       for (String value in valuesBySendType[sendType].keys) {
-        chartDataMap[sendType].add(AttemptsByValueSeries(value, valuesBySendType[sendType][value]));
+        chartDataMap[sendType].add(
+            AttemptsByValueSeries(value, valuesBySendType[sendType][value]));
       }
     }
 
@@ -162,7 +194,8 @@ class _AttemptsByValueChartState extends State<AttemptsByValueChart> {
     SendTypes.SEND_TYPES.asMap().forEach((index, sendType) {
       chartSeries.add(charts.Series<AttemptsByValueSeries, String>(
         id: sendType,
-        colorFn: (_, __) => charts.ColorUtil.fromDartColor(SeriesConstants.COLOURS[index]),
+        colorFn: (_, __) =>
+            charts.ColorUtil.fromDartColor(SeriesConstants.COLOURS[index]),
         domainFn: (AttemptsByValueSeries attempts, _) => attempts.value,
         measureFn: (AttemptsByValueSeries attempts, _) => attempts.count,
         data: chartDataMap[sendType],