+++ lib/stats/charts/downclimbed_by_grade.dart
@@ -11,11 +11,11 @@ import 'attempt_filter.dart';
 class DownclimbedByGradeChart extends StatefulWidget {
   DownclimbedByGradeChart(
       {Key key,
-        @required this.attempts,
-        @required this.categories,
-        @required this.grades,
-        @required this.locationNamesToIds,
-        @required this.locationNamesToGradeSet})
+      @required this.attempts,
+      @required this.categories,
+      @required this.grades,
+      @required this.locationNamesToIds,
+      @required this.locationNamesToGradeSet})
       : super(key: key);
   final List<Attempt> attempts;
   final List<String> categories;
@@ -23,7 +23,8 @@ class DownclimbedByGradeChart extends StatefulWidget {
   final Map<String, String> locationNamesToIds;
   final Map<String, String> locationNamesToGradeSet;
   @override
-  _DownclimbedByGradeChartState createState() => _DownclimbedByGradeChartState();
+  _DownclimbedByGradeChartState createState() =>
+      _DownclimbedByGradeChartState();
 }
 
 class _DownclimbedByGradeChartState extends State<DownclimbedByGradeChart> {
@@ -37,7 +38,8 @@ class _DownclimbedByGradeChartState extends State<DownclimbedByGradeChart> {
   @override
   void initState() {
     filteredAttemptsStream = StreamController<List<Attempt>>.broadcast();
-    filteredAttemptsListener = filteredAttemptsStream.stream.listen((filteredAttempts) {
+    filteredAttemptsListener =
+        filteredAttemptsStream.stream.listen((filteredAttempts) {
       setState(() {
         if (filterGradeSet != null) {
           chartSeries = _buildChartSeries(context, filteredAttempts);
@@ -90,38 +92,64 @@ class _DownclimbedByGradeChartState extends State<DownclimbedByGradeChart> {
           primaryMeasureAxis: charts.NumericAxisSpec(
               renderSpec: charts.GridlineRendererSpec(
                 labelStyle: charts.TextStyleSpec(
-                  fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
-                  fontWeight: Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
-                  color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
+                  fontSize: Theme.of(context)
+                      .accentTextTheme
+                      .caption
+                      .fontSize
+                      .toInt(),
+                  fontWeight: Theme.of(context)
+                      .accentTextTheme
+                      .caption
+                      .fontWeight
+                      .toString(),
+                  color: charts.ColorUtil.fromDartColor(
+                      Theme.of(context).accentColor),
                 ),
                 lineStyle: charts.LineStyleSpec(
-                  color: charts.ColorUtil.fromDartColor(Theme.of(context).dialogBackgroundColor),
+                  color: charts.ColorUtil.fromDartColor(
+                      Theme.of(context).dialogBackgroundColor),
                 ),
               ),
-              tickProviderSpec: charts.StaticNumericTickProviderSpec(_buildMeasureTicks()),
-              tickFormatterSpec: charts.BasicNumericTickFormatterSpec.fromNumberFormat(
-                  NumberFormat.percentPattern())),
+              tickProviderSpec:
+                  charts.StaticNumericTickProviderSpec(_buildMeasureTicks()),
+              tickFormatterSpec:
+                  charts.BasicNumericTickFormatterSpec.fromNumberFormat(
+                      NumberFormat.percentPattern())),
           domainAxis: charts.OrdinalAxisSpec(
             renderSpec: charts.SmallTickRendererSpec(
               lineStyle: charts.LineStyleSpec(
-                color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
+                color: charts.ColorUtil.fromDartColor(
+                    Theme.of(context).accentColor),
               ),
               labelStyle: charts.TextStyleSpec(
-                fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
-                fontWeight: Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
-                color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
+                fontSize:
+                    Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
+                fontWeight: Theme.of(context)
+                    .accentTextTheme
+                    .caption
+                    .fontWeight
+                    .toString(),
+                color: charts.ColorUtil.fromDartColor(
+                    Theme.of(context).accentColor),
               ),
             ),
-            tickProviderSpec: charts.StaticOrdinalTickProviderSpec(_buildDomainTicks()),
+            tickProviderSpec:
+                charts.StaticOrdinalTickProviderSpec(_buildDomainTicks()),
           ),
           barGroupingType: charts.BarGroupingType.grouped,
           defaultInteractions: false,
           behaviors: [
             charts.SeriesLegend(
               entryTextStyle: charts.TextStyleSpec(
-                fontSize: Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
-                fontWeight: Theme.of(context).accentTextTheme.caption.fontWeight.toString(),
-                color: charts.ColorUtil.fromDartColor(Theme.of(context).accentColor),
+                fontSize:
+                    Theme.of(context).accentTextTheme.caption.fontSize.toInt(),
+                fontWeight: Theme.of(context)
+                    .accentTextTheme
+                    .caption
+                    .fontWeight
+                    .toString(),
+                color: charts.ColorUtil.fromDartColor(
+                    Theme.of(context).accentColor),
               ),
             ),
           ],
@@ -129,7 +157,7 @@ class _DownclimbedByGradeChartState extends State<DownclimbedByGradeChart> {
       } else {
         content = Center(
             child: Text(
-          "There are no existing attempts ${widget.attempts.isNotEmpty ? "matching these filters" : ""}. \nGo log some!",
+          "There are no existing attempts${widget.attempts.isNotEmpty ? " matching these filters" : ""}. \nGo log some!",
           style: Theme.of(context).accentTextTheme.subtitle2,
           textAlign: TextAlign.center,
         ));
@@ -148,7 +176,8 @@ class _DownclimbedByGradeChartState extends State<DownclimbedByGradeChart> {
             padding: EdgeInsets.all(UIConstants.SMALLER_PADDING),
             decoration: BoxDecoration(
                 color: Theme.of(context).cardColor,
-                borderRadius: BorderRadius.all(Radius.circular(UIConstants.CARD_BORDER_RADIUS))),
+                borderRadius: BorderRadius.all(
+                    Radius.circular(UIConstants.CARD_BORDER_RADIUS))),
             child: content));
   }
 
@@ -160,7 +189,8 @@ class _DownclimbedByGradeChartState extends State<DownclimbedByGradeChart> {
 
     List<String> sendTypes = List.from(SendTypes.SEND_TYPES)..removeLast();
 
-    Map<String, Map<String, List<int>>> attemptsBySendType = <String, Map<String, List<int>>>{};
+    Map<String, Map<String, List<int>>> attemptsBySendType =
+        <String, Map<String, List<int>>>{};
     for (String sendType in sendTypes) {
       attemptsBySendType.putIfAbsent(
           sendType,
@@ -171,8 +201,10 @@ class _DownclimbedByGradeChartState extends State<DownclimbedByGradeChart> {
     for (Attempt attempt in filteredAttempts) {
       String grade = attempt.climbGrade;
       if (attempt.sendType != "Attempt") {
-        attemptsBySendType[attempt.sendType]
-            .update(grade, (counts) => [counts[0] + (attempt.downclimbed ? 1 : 0), counts[1] + 1]);
+        attemptsBySendType[attempt.sendType].update(
+            grade,
+            (counts) =>
+                [counts[0] + (attempt.downclimbed ? 1 : 0), counts[1] + 1]);
       }
     }
 
@@ -195,7 +227,8 @@ class _DownclimbedByGradeChartState extends State<DownclimbedByGradeChart> {
     sendTypes.asMap().forEach((index, sendType) {
       chartSeries.add(charts.Series<DownclimbedByGradeSeries, String>(
         id: sendType,
-        colorFn: (_, __) => charts.ColorUtil.fromDartColor(SeriesConstants.COLOURS[index]),
+        colorFn: (_, __) =>
+            charts.ColorUtil.fromDartColor(SeriesConstants.COLOURS[index]),
         domainFn: (DownclimbedByGradeSeries attempts, _) => attempts.grade,
         measureFn: (DownclimbedByGradeSeries attempts, _) =>
             (attempts.count == 0) ? 0.0 : attempts.downclimbed / attempts.count,
@@ -216,7 +249,9 @@ class _DownclimbedByGradeChartState extends State<DownclimbedByGradeChart> {
   }
 
   List<charts.TickSpec<String>> _buildDomainTicks() {
-    return widget.grades[filterGradeSet].map((grade) => charts.TickSpec(grade)).toList();
+    return widget.grades[filterGradeSet]
+        .map((grade) => charts.TickSpec(grade))
+        .toList();
   }
 
   @override