+++ lib/view_only_climb/view_only_climb_view.dart
@@ -72,6 +72,8 @@ class ViewOnlyClimbWidget extends StatelessWidget {
             content = Center(
               child: CircularProgressIndicator(
                 strokeWidth: 4.0,
+                valueColor: AlwaysStoppedAnimation<Color>(
+                    Theme.of(context).accentColor),
               ),
             );
           } else {
@@ -173,6 +175,8 @@ class ViewOnlyClimbWidget extends StatelessWidget {
             child: Center(
                 child: CircularProgressIndicator(
               strokeWidth: 4.0,
+              valueColor:
+                  AlwaysStoppedAnimation<Color>(Theme.of(context).accentColor),
             ))),
         errorWidget: (context, url, error) => Icon(Icons.error),
       ),